Move emixx mixins into their own package
