+++ common/src/main/resources/emi_discovery.mixins.json
@@ -5,13 +5,13 @@
     "refmap": "${mod_id}.refmap.json",
     "compatibilityLevel": "JAVA_17",
     "mixins": [
-        "EMIxxStackManagerAccessor",
+        "emixx.EMIxxStackManagerAccessor",
         "MinecraftServerStorageSourceAccessor",
         ,
         "emi.accessor.PageSlotWidgetAccessor"
     ],
     "client": [
-        "EMIxxStackManagerMixin",
+        "emixx.EMIxxStackManagerMixin",
         "emi.EmiScreenManagerMixin",
         "emi.EmiSidebarsMixin",
         "emi.PageSlotWidgetMixin",
