+++ fabric/src/main/resources/emi_discovery.fabric.mixins.json
@@ -6,7 +6,7 @@
     "compatibilityLevel": "JAVA_17",
     "mixins": [
         "InventoryChangeTriggerMixinFabric",
-        "EMIxxStackManagerMixin"
+        "emixx.EMIxxStackManagerMixin"
     ],
     "client": [
         "ClientPacketListenerMixinFabric",
